Removed lot of tone listener code from sipxcallib. There was lot of junk that was never used. Removed old DTMF RFC2833 notification system from sipxcalllib.
New MpResNotification was added to sipxmedialib. It is meant to notify resources of any action when direct knowledge of another resource is unwanted. It is currently used for MpdPtAVT rfc2833 DTMF decoder to notify it's parent MprDecode about DTMF.
In the future, also other resources might use this system to avoid circular dependencies.

Code which informed recorders about certain DTMF event was removed from sipxmedialib. Recording should be stopped manually from sipxtapi.

Both in-band and RFC2833 DTMF now use new notification code. In-band decoder was tested successfuly with 3 different phones, only SJPhone in-band detection doesn't work properly, probably due to a bug in SJPhone.

Enabling/Disabling DTMF decoders is not possible yet.
All DTMF decoders are enabled by default.

+++ trunk/sipXmediaLib/src/mp/MpAudioResource.cpp
@@ -232,7 +232,7 @@ int MpAudioResource::getSamplesPerSec()
    return mSamplesPerSec;
 }
 
-void MpAudioResource::sendInterfaceNotification(MpNotificationMsgType type, int data)
+void MpAudioResource::sendInterfaceNotification(MpNotificationMsgType type, intptr_t data)
 {
    if (mpFlowGraph && areNotificationsEnabled())
    {
